fix(ci): 修复大文件与重测试噪音门禁
