+++ package.json
@@ -35,7 +35,7 @@
     "perf:realtime:boundary-guard": "vitest run src/features/threads/contracts/realtimeBoundaryGuard.test.ts",
     "lint": "eslint . --ext .ts,.tsx",
     "test": "node scripts/test-batched.mjs",
-    "test:integration": "VITEST_INCLUDE_HEAVY=1 node scripts/test-batched.mjs",
+    "test:integration": "node scripts/test-batched.mjs --include-heavy",
     "test:layout-guard": "vitest run src/styles/layout-swapped-platform-guard.test.ts",
     "test:memory-kind-contract:ts": "vitest run src/features/project-memory/utils/memoryKindClassifier.contract.test.ts",
     "test:memory-kind-contract:rust": "cargo test --manifest-path src-tauri/Cargo.toml classify_kind_matches_contract_samples",
