+++ scripts/test-batched.mjs
@@ -1,64 +1,85 @@
 import { execFileSync, spawnSync } from "node:child_process";
 import { readdirSync } from "node:fs";
-import { join } from "node:path";
+import { join, resolve } from "node:path";
+import { pathToFileURL } from "node:url";
 
-const batchSize = Number.parseInt(process.env.VITEST_BATCH_SIZE ?? "4", 10);
-const includeHeavyIntegration = process.env.VITEST_INCLUDE_HEAVY === "1";
-if (!Number.isFinite(batchSize) || batchSize <= 0) {
-  throw new Error(`Invalid VITEST_BATCH_SIZE: ${process.env.VITEST_BATCH_SIZE ?? ""}`);
-}
-
-const testFiles = listTestFiles().filter((file) => {
-  if (includeHeavyIntegration) {
-    return true;
+export function parseVitestBatchConfig(argv = [], env = process.env) {
+  let includeHeavyIntegration = env.VITEST_INCLUDE_HEAVY === "1";
+  for (const token of argv) {
+    if (token === "--include-heavy") {
+      includeHeavyIntegration = true;
+      continue;
+    }
+    throw new Error(`Unknown argument: ${token}`);
   }
-  return !file.endsWith(".integration.test.tsx");
-});
 
-if (testFiles.length === 0) {
-  console.log("No test files found.");
-  process.exit(0);
+  const batchSize = Number.parseInt(env.VITEST_BATCH_SIZE ?? "4", 10);
+  if (!Number.isFinite(batchSize) || batchSize <= 0) {
+    throw new Error(`Invalid VITEST_BATCH_SIZE: ${env.VITEST_BATCH_SIZE ?? ""}`);
+  }
+
+  return {
+    batchSize,
+    includeHeavyIntegration,
+  };
 }
 
-if (!includeHeavyIntegration) {
-  console.log(
-    "[vitest-batch] heavy *.integration.test.tsx suites are excluded by default; set VITEST_INCLUDE_HEAVY=1 to include.",
-  );
-}
+export function runVitestBatches(argv = process.argv.slice(2), env = process.env) {
+  const { batchSize, includeHeavyIntegration } = parseVitestBatchConfig(argv, env);
+  const testFiles = listTestFiles().filter((file) => {
+    if (includeHeavyIntegration) {
+      return true;
+    }
+    return !file.endsWith(".integration.test.tsx");
+  });
 
-const totalBatches = Math.ceil(testFiles.length / batchSize);
-for (let i = 0; i < totalBatches; i += 1) {
-  const start = i * batchSize;
-  const end = Math.min(start + batchSize, testFiles.length);
-  const files = testFiles.slice(start, end);
-  console.log(
-    `[vitest-batch] ${i + 1}/${totalBatches} files ${start + 1}-${end}/${testFiles.length}`,
-  );
-  const result = spawnSync(
-    process.execPath,
-    [
-      "node_modules/vitest/vitest.mjs",
-      "run",
-      "--maxWorkers",
-      "1",
-      "--minWorkers",
-      "1",
-      ...files,
-    ],
-    {
-      stdio: "inherit",
-      env: {
-        ...process.env,
-        NODE_OPTIONS: withMemoryOption(process.env.NODE_OPTIONS, 12288),
+  if (testFiles.length === 0) {
+    console.log("No test files found.");
+    return 0;
+  }
+
+  if (!includeHeavyIntegration) {
+    console.log(
+      "[vitest-batch] heavy *.integration.test.tsx suites are excluded by default; set VITEST_INCLUDE_HEAVY=1 or pass --include-heavy to include.",
+    );
+  }
+
+  const totalBatches = Math.ceil(testFiles.length / batchSize);
+  for (let i = 0; i < totalBatches; i += 1) {
+    const start = i * batchSize;
+    const end = Math.min(start + batchSize, testFiles.length);
+    const files = testFiles.slice(start, end);
+    console.log(
+      `[vitest-batch] ${i + 1}/${totalBatches} files ${start + 1}-${end}/${testFiles.length}`,
+    );
+    const result = spawnSync(
+      process.execPath,
+      [
+        "node_modules/vitest/vitest.mjs",
+        "run",
+        "--maxWorkers",
+        "1",
+        "--minWorkers",
+        "1",
+        ...files,
+      ],
+      {
+        stdio: "inherit",
+        env: {
+          ...env,
+          VITEST_INCLUDE_HEAVY: includeHeavyIntegration ? "1" : env.VITEST_INCLUDE_HEAVY,
+          NODE_OPTIONS: withMemoryOption(env.NODE_OPTIONS, 12288),
+        },
       },
-    },
-  );
-  if (result.status !== 0) {
-    process.exit(result.status ?? 1);
+    );
+    if (result.status !== 0) {
+      return result.status ?? 1;
+    }
   }
-}
 
-console.log(`[vitest-batch] completed ${testFiles.length} test files.`);
+  console.log(`[vitest-batch] completed ${testFiles.length} test files.`);
+  return 0;
+}
 
 function listTestFiles() {
   const fromRipgrep = listWithRipgrep();
@@ -121,4 +142,12 @@ function withMemoryOption(currentOptions, memoryMb = 12288) {
     return currentOptions;
   }
   return `${currentOptions} ${option}`.trim();
+}
+
+const isDirectExecution =
+  process.argv[1] &&
+  import.meta.url === pathToFileURL(resolve(process.argv[1])).href;
+
+if (isDirectExecution) {
+  process.exit(runVitestBatches());
 }
