+++ src/i18n/locales/en.part1.ts
@@ -2870,136 +2870,6 @@ const enPart1 = {
     },
   },
 
-  memory: {
-    title: "Project Memory",
-    settings: "Memory settings",
-    loading: "Loading memories...",
-    empty: "No memory records yet.",
-    selectWorkspace: "Select a workspace to view memory.",
-    selectRecord: "Select a record to view details.",
-    searchPlaceholder: "Search memory...",
-    autoCaptureWorkspace: "Enable auto capture for this workspace",
-    contextInjectionEnabled: "Enable memory context injection for chat",
-    contextInjectionManualHint: "Switched to manual memory linking with @@ in chat (one-shot injection).",
-    manualInjectionMode: "Manual injection strategy",
-    manualInjectionModeDetail: "Detail mode (detail)",
-    manualInjectionModeSummary: "Summary mode (summary)",
-    manualInjectionModeHint: "Default is detail mode: inject memory detail, while canvas still shows summary only.",
-    cleanupTitle: "Polluted memory cleanup",
-    cleanupScan: "Scan polluted memory",
-    cleanupScanning: "Scanning...",
-    cleanupRun: "Run cleanup",
-    cleanupRunning: "Cleaning...",
-    cleanupHint: "Scan first, verify matches, then delete.",
-    cleanupScanned: "Scanned {{total}} memories.",
-    cleanupPreview: "Scan completed: {{matched}} matches out of {{scanned}} scanned.",
-    cleanupResult: "Cleanup completed: deleted {{success}}, failed {{failed}}.",
-    refresh: "Refresh",
-    openManager: "Open Manager",
-    closeManager: "Close Manager",
-    help: "Help",
-    closeHelp: "Close help",
-    save: "Save",
-    saving: "Saving...",
-    delete: "Delete",
-    add: "Add memory",
-    newTitle: "Title (optional)",
-    newDetail: "Memory detail",
-    detailTagsPlaceholder: "Tags (comma separated)",
-    detailPreviewTitle: "Formatted preview",
-    detailPreviewEmpty: "No content to preview.",
-    helpTitle: "Panel Guide",
-    helpButtons: "Top buttons: 🔄 Refresh memory list · ⚙️ Open settings (auto-capture, context injection) · ❓ Open help · ✕ Close panel",
-    helpFilters: "Toolbar: Search keywords · Filter by kind · Filter by priority · Filter by tag",
-    helpKind: "Kind tag: Memory category — Project context / Conversation / Code decision / Known issue / Note",
-    helpImportance: "Priority tag: Memory importance — High / Medium / Low, used for filtering and sorting",
-    helpEngine: "Engine tag: Conversation-only, indicates which AI model generated the memory",
-    helpAutoCapture: "Auto capture: When enabled, important information from conversations is automatically saved as project memory",
-    helpContextInjection: "Context injection: When enabled, relevant memories are automatically injected into conversation context to improve AI response accuracy",
-    helpBatchOps: "Batch operations: Check list items to batch delete selected memories; Settings area allows clearing all memories",
-    helpMemoryTypes:
-      "Memory types: Project context (long-term project info), Conversation (AI chat records), Code decision (architecture decisions), Known issue (bugs and pending issues), Note (manually added custom memories)",
-    selectAll: "Select all",
-    unselectAll: "Unselect all",
-    batchDelete: "Delete selected",
-    batchSetHigh: "Set selected to High",
-    batchSetMedium: "Set selected to Medium",
-    batchSetLow: "Set selected to Low",
-    batchDeleteConfirm: "Delete {{count}} selected memories? This action cannot be undone.",
-    clearAll: "Clear all memories",
-    clearAllConfirm: "Clear all memories? This action cannot be undone!",
-    clearAllSuccess: "Cleared {{count}} memories",
-    batchDeleteSuccess: "Deleted {{count}} memories",
-    batchUpdateImportanceSuccess: "Updated {{count}} memories to {{importance}} priority",
-    selected: "{{count}} selected",
-    cancel: "Cancel",
-    confirmDelete: "Confirm Delete",
-    tagPlaceholder: "Filter by tag",
-    quickTags: "Quick tags",
-    prevPage: "Prev",
-    nextPage: "Next",
-    pageMeta: "{{from}}-{{to}} / {{total}}",
-    importance: {
-      all: "All priorities",
-      high: "High",
-      medium: "Medium",
-      low: "Low",
-    },
-    kind: {
-      all: "All kinds",
-      projectContext: "Project context",
-      conversation: "Conversation",
-      codeDecision: "Code decision",
-      knownIssue: "Known issue",
-      note: "Note",
-    },
-    checking: "Connecting to long-term memory service…",
-    offlineTitle: "Long-term Memory Service Unavailable",
-    offlineDesc: "Long-term memory is powered by the open-source project claude-mem. Follow the steps below to install it.",
-    offlineStepsTitle: "Open a new Claude Code session in your terminal and run:",
-    offlineStep1: "/plugin marketplace add thedotmack/claude-mem",
-    offlineStep2: "/plugin install claude-mem",
-    offlineStep3: "Restart Claude Code in terminal",
-    offlineStep4: "Restart ccgui",
-  },
-
-  // Time
-  time: {
-    justNow: "just now",
-    minuteAgo: "{{count}} minute ago",
-    minutesAgo: "{{count}} minutes ago",
-    hourAgo: "{{count}} hour ago",
-    hoursAgo: "{{count}} hours ago",
-    dayAgo: "{{count}} day ago",
-    daysAgo: "{{count}} days ago",
-    weekAgo: "{{count}} week ago",
-    weeksAgo: "{{count}} weeks ago",
-    // Short relative time formats
-    now: "now",
-    shortSecondsAgo: "{{value}}s ago",
-    shortSecondsIn: "in {{value}}s",
-    shortMinutesAgo: "{{value}}m ago",
-    shortMinutesIn: "in {{value}}m",
-    shortMinutes: "{{value}}m",
-    shortHours: "{{value}}h",
-    shortDays: "{{value}}d",
-    shortWeeks: "{{value}}w",
-    shortMonths: "{{value}}mo",
-    shortYears: "{{value}}y",
-  },
-
-  // About
-  about: {
-    version: "Version",
-    tagline: "Next-generation VibeCoding editor",
-    checkForUpdates: "Check for Updates",
-    releaseNotes: "Release Notes",
-    reportIssue: "Report Issue",
-    website: "Website",
-    github: "GitHub",
-    wechatGroupTitle: "Scan to join WeChat Group",
-  },
-
 };
 
 export default enPart1;
