+++ src/i18n/locales/en.ts
@@ -1,6 +1,7 @@
 import enPart1 from "./en.part1";
 import enPart2 from "./en.part2";
 import enPart3 from "./en.part3";
+import enPart4 from "./en.part4";
 
 const enPart2Settings = (enPart2 as { settings?: Partial<typeof enPart1.settings> }).settings ?? {};
 const enPart3Settings = (enPart3 as { settings?: Partial<typeof enPart1.settings> }).settings ?? {};
@@ -11,6 +12,7 @@ const en = {
   ...enPart1,
   ...enPart2,
   ...enPart3,
+  ...enPart4,
   composer: {
     ...enPart1.composer,
     ...enPart2Composer,
